document ExperimentDef
